TensorFlow/TF1/VisionTransformer/requirements.txt
tensorflow-directml
sklearn
numpy
